lambdas/textract_cracking/src/requirements.txt
unstructured
psycopg2-binary
pgvector
boto3
amazon-textract-textractor
amazon-textract-textractor[pdf]